feat(db): migrate remaining SQLite stores to Postgres (closes #747)
Migrates event_store, review_store, thread_db, project_registry,
runtime_state_store, q_value_store, and plan_db from SQLite to
PostgreSQL following the task_db pattern (pg_open_pool_schematized +
PgMigrator). Large files split into directory modules to stay under the
800-line limit. Pre-commit hook excludes harness-observe and
harness-workflow from lib tests when DATABASE_URL is absent.
