+++ .githooks/pre-commit
@@ -18,8 +18,8 @@ echo "=== pre-commit: test ==="
 if [ -n "${DATABASE_URL:-}" ]; then
     cargo test --workspace --lib
 else
-    echo "DATABASE_URL not set — skipping harness-server lib tests (need Postgres)"
-    cargo test --workspace --lib --exclude harness-server
+    echo "DATABASE_URL not set — skipping Postgres-dependent lib tests (harness-server, harness-observe, harness-workflow)"
+    cargo test --workspace --lib --exclude harness-server --exclude harness-observe --exclude harness-workflow
 fi
 
 echo "=== pre-commit: all checks passed ==="
